Introduce hoverDelegate2.ts
Part of #204965
